# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Wehda Bus Booking API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Trip {
  id: number;
  from: string;
  to: string;
  fromStation?: string;
  toStation?: string;
  departure: string;
  arrival: string;
  duration: string;
  busType: string;
  price: number;
  seatsLeft: number;
  totalSeats: number;
  status: string;
  amenities: string[];
}

export type SeatState = typeof SeatState[keyof typeof SeatState];


export const SeatState = {
  available: 'available',
  booked: 'booked',
  held: 'held',
  disabled: 'disabled',
} as const;

export interface Seat {
  number: number;
  state: SeatState;
  /** @nullable */
  passengerName?: string | null;
}

export interface SeatMap {
  tripId: number;
  rows: number;
  seats: Seat[];
}

export interface BookingHoldInput {
  tripId: number;
  /**
     * @minItems 1
     * @maxItems 4
     */
  seats: number[];
  /** @minLength 2 */
  passengerName: string;
  /** @pattern ^01[0-9]{9}$ */
  passengerPhone: string;
  pickupType: string;
  /** @nullable */
  pickupPoint?: string | null;
  waitMinutes?: number;
}

export type PaymentInputMethod = typeof PaymentInputMethod[keyof typeof PaymentInputMethod];


export const PaymentInputMethod = {
  vodafone_cash: 'vodafone_cash',
  instapay: 'instapay',
} as const;

export interface PaymentInput {
  method: PaymentInputMethod;
  /** @minLength 3 */
  transactionRef: string;
  senderPhone: string;
}

export type ReviewInputDecision = typeof ReviewInputDecision[keyof typeof ReviewInputDecision];


export const ReviewInputDecision = {
  approve: 'approve',
  reject: 'reject',
} as const;

export interface ReviewInput {
  decision: ReviewInputDecision;
  /** @nullable */
  reason?: string | null;
}

export type BookingCustomer = {
  name: string;
  phone: string;
};

export type BookingPassengersItem = {
  name: string;
  seat: number;
};

export interface Booking {
  id: number;
  reference: string;
  status: string;
  paymentStatus: string;
  total: number;
  pickupFee?: number;
  pickupType?: string;
  createdAt?: string;
  expiresAt?: string;
  seats: number[];
  trip: Trip;
  customer: BookingCustomer;
  passengers?: BookingPassengersItem[];
}

export type StaffDashboardAttentionItem = {
  title: string;
  subtitle: string;
  tone: string;
};

export type StaffDashboardRevenueItem = {
  label: string;
  value: number;
};

export interface StaffDashboard {
  todayTrips: number;
  newBookings: number;
  pendingPayments: number;
  pendingPickups: number;
  freeSeats: number;
  todayRevenue: number;
  attention: StaffDashboardAttentionItem[];
  revenue: StaffDashboardRevenueItem[];
}

export type FromParameter = string;

export type ToParameter = string;

export type DateParameter = string;

export type ListTripsParams = {
from?: FromParameter;
to?: ToParameter;
date?: DateParameter;
/**
 * @minimum 1
 * @maximum 4
 */
passengers?: number;
};

export type ListStaffBookingsParams = {
status?: string;
search?: string;
};