# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Wehda Bus Booking API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingCustomer } from './bookingCustomer';
import type { BookingPassengersItem } from './bookingPassengersItem';
import type { Trip } from './trip';

export interface Booking {
  id: number;
  reference: string;
  status: string;
  paymentStatus: string;
  total: number;
  pickupFee?: number;
  pickupType?: string;
  createdAt?: string;
  expiresAt?: string;
  seats: number[];
  trip: Trip;
  customer: BookingCustomer;
  passengers?: BookingPassengersItem[];
}