# lib/api-zod/src/generated/types/bookingHoldInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Wehda Bus Booking API
 * OpenAPI spec version: 0.1.0
 */

export interface BookingHoldInput {
  tripId: number;
  /**
     * @minItems 1
     * @maxItems 4
     */
  seats: number[];
  /** @minLength 2 */
  passengerName: string;
  /** @pattern ^01[0-9]{9}$ */
  passengerPhone: string;
  pickupType: string;
  /** @nullable */
  pickupPoint?: string | null;
  waitMinutes?: number;
}