# lib/api-zod/src/generated/types/paymentInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Wehda Bus Booking API
 * OpenAPI spec version: 0.1.0
 */
import type { PaymentInputMethod } from './paymentInputMethod';

export interface PaymentInput {
  method: PaymentInputMethod;
  /** @minLength 3 */
  transactionRef: string;
  senderPhone: string;
}