# lib/api-zod/src/generated/types/reviewInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Wehda Bus Booking API
 * OpenAPI spec version: 0.1.0
 */
import type { ReviewInputDecision } from './reviewInputDecision';

export interface ReviewInput {
  decision: ReviewInputDecision;
  /** @nullable */
  reason?: string | null;
}