# lib/api-zod/src/generated/types/seat.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Wehda Bus Booking API
 * OpenAPI spec version: 0.1.0
 */
import type { SeatState } from './seatState';

export interface Seat {
  number: number;
  state: SeatState;
  /** @nullable */
  passengerName?: string | null;
}