# lib/api-zod/src/generated/types/seatMap.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Wehda Bus Booking API
 * OpenAPI spec version: 0.1.0
 */
import type { Seat } from './seat';

export interface SeatMap {
  tripId: number;
  rows: number;
  seats: Seat[];
}