# lib/api-zod/src/generated/types/staffDashboard.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Wehda Bus Booking API
 * OpenAPI spec version: 0.1.0
 */
import type { StaffDashboardAttentionItem } from './staffDashboardAttentionItem';
import type { StaffDashboardRevenueItem } from './staffDashboardRevenueItem';

export interface StaffDashboard {
  todayTrips: number;
  newBookings: number;
  pendingPayments: number;
  pendingPickups: number;
  freeSeats: number;
  todayRevenue: number;
  attention: StaffDashboardAttentionItem[];
  revenue: StaffDashboardRevenueItem[];
}