# lib/api-zod/src/generated/types/trip.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Wehda Bus Booking API
 * OpenAPI spec version: 0.1.0
 */

export interface Trip {
  id: number;
  from: string;
  to: string;
  fromStation?: string;
  toStation?: string;
  departure: string;
  arrival: string;
  duration: string;
  busType: string;
  price: number;
  seatsLeft: number;
  totalSeats: number;
  status: string;
  amenities: string[];
}